# backend/docs/swagger.yaml
basePath: /api/
definitions:
...
  database.Player:
    properties:
      group_id:
        type: integer
      id:
        type: integer
      lane_id:
        type: integer
      name:
        type: string
      order:
        type: integer
      participant_id:
        type: integer
      player_sets:
        items:
          $ref: '#/definitions/database.PlayerSet'
        type: array
      rank:
        type: integer
      rounds:
        items:
          $ref: '#/definitions/database.Round'
        type: array
      shoot_off_score:
        type: integer
      total_score:
        type: integer
    type: object
  database.PlayerSet:
    properties:
      elimination_id:
        type: integer
      id:
        type: integer
      players:
        items:
          $ref: '#/definitions/database.Player'
        type: array
      rank:
        type: integer
      set_name:
        type: string
      total_score:
        type: integer
    type: object
...
  database.Round:
    properties:
      id:
        type: integer
      player_id:
        type: integer
      round_ends:
        items:
          $ref: '#/definitions/database.RoundEnd'
        type: array
      total_score:
        type: integer
    type: object
  database.RoundEnd:
    properties:
      id:
        type: integer
      is_confirmed:
        type: boolean
      round_id:
        type: integer
      round_scores:
        items:
          $ref: '#/definitions/database.RoundScore'
        type: array
    type: object
  database.RoundScore:
    properties:
      id:
        type: integer
      round_end_id:
        type: integer
      score:
        type: integer
    type: object
...
  endpoint.GetPlayerSetsByMedalByEliminationId.playerSetData:
    properties:
      id:
        type: integer
      set_name:
        type: string
      type:
        type: integer
    type: object
...
  endpoint.PostPlayerSet.playerSetData:
    properties:
      elimination_id:
        type: integer
      player_ids:
        items:
          type: integer
        type: array
      set_name:
        type: string
    type: object
...
  endpoint.PutPlayerSetName.playerSetData:
    properties:
      set_name:
        type: string
    type: object
...
  response.DeleteSuccessResponse:
    properties:
      message:
        example: 'Delete ID(1) : sth delete success'
        type: string
    type: object
  response.ErrorIdResponse:
    properties:
      error:
        example: 'invalid ID(1) : sth error'
        type: string
    type: object
  response.ErrorInternalErrorResponse:
    properties:
      error:
        example: 'sth need fix ID(1) : sth error'
        type: string
    type: object
  response.ErrorReceiveDataFormatResponse:
    properties:
      error:
        example: 'bad request data: sth error'
        type: string
    type: object
...
  response.Nill:
    type: object
...
host: localhost:80
info:
  contact:
    name: NYCUArchery
    url: https://github.com/NYCUarchery
  description: Gin swagger
  license:
    name: no license yet
  title: Gin swagger
  version: "1.0"
paths:
...
  /playerset:
    post:
      consumes:
      - application/json
      description: |-
        Post player set, and build player set match table
        If team size is 1, set name will be player name
      parameters:
      - description: Player Set Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/endpoint.PostPlayerSet.playerSetData'
      produces:
      - application/json
      responses:
        "200":
          description: success, return player set without players
          schema:
            allOf:
            - $ref: '#/definitions/database.PlayerSet'
            - properties:
                players:
                  $ref: '#/definitions/response.Nill'
              type: object
        "400":
          description: invalid elimination id, player id maybe not exist / player
            set's length and teamsize does not match
          schema:
            $ref: '#/definitions/response.ErrorIdResponse'
        "500":
          description: internal db error for create player set / get player / create
            player set match table / get elimination
          schema:
            $ref: '#/definitions/response.ErrorInternalErrorResponse'
      summary: Post player set
      tags:
      - PlayerSet
  /playerset/{id}:
    delete:
      description: Delete player set, and delete player set match table
      parameters:
      - description: Player Set ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/response.DeleteSuccessResponse'
        "400":
          description: invalid plyer set id
          schema:
            $ref: '#/definitions/response.ErrorIdResponse'
        "500":
          description: internal db error / Get Player Set By Id / Delete Player Set
            Match Table By Player Set Id / Delete Player Set By Id
          schema:
            $ref: '#/definitions/response.ErrorInternalErrorResponse'
      summary: Delete player set
      tags:
      - PlayerSet
    get:
      description: Get player set with players by id
      parameters:
      - description: Player Set ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success, return player set with players
          schema:
            allOf:
            - $ref: '#/definitions/database.PlayerSet'
            - properties:
                players:
                  allOf:
                  - $ref: '#/definitions/database.Player'
                  - properties:
                      player_sets:
                        $ref: '#/definitions/response.Nill'
                      rounds:
                        $ref: '#/definitions/response.Nill'
                    type: object
              type: object
        "400":
          description: invalid player set id, maybe not exist
          schema:
            $ref: '#/definitions/response.ErrorIdResponse'
        "500":
          description: internal error for get player set by id
          schema:
            $ref: '#/definitions/response.ErrorInternalErrorResponse'
      summary: Get player set with players by id
      tags:
      - PlayerSet
  /playerset/elimination/{eliminationid}:
    get:
      description: Get all player sets by elimination id
      parameters:
      - description: Elimination ID
        in: path
        name: eliminationid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            items:
              $ref: '#/definitions/database.PlayerSet'
            type: array
        "400":
          description: invalid player set id
          schema:
            $ref: '#/definitions/response.ErrorIdResponse'
        "500":
          description: internal db error / Get Player Sets By Elimination Id
          schema:
            $ref: '#/definitions/response.ErrorInternalErrorResponse'
      summary: Get all player sets by elimination id
      tags:
      - PlayerSet
  /playerset/elimination/medal/{eliminationid}:
    get:
      description: Get player sets which have medals by elimination id
      parameters:
      - description: Elimination ID
        in: path
        name: eliminationid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            items:
              $ref: '#/definitions/endpoint.GetPlayerSetsByMedalByEliminationId.playerSetData'
            type: array
        "400":
          description: invalid elimination id
          schema:
            $ref: '#/definitions/response.ErrorIdResponse'
        "500":
          description: internal db error / Get Medal Info By Elimination Id
          schema:
            $ref: '#/definitions/response.ErrorInternalErrorResponse'
      summary: Get player sets which have medals by elimination id
      tags:
      - PlayerSet
  /playerset/name/{id}:
    patch:
      consumes:
      - application/json
      description: Put player set name
      parameters:
      - description: Player Set ID
        in: path
        name: id
        required: true
        type: integer
      - description: Player Set Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/endpoint.PutPlayerSetName.playerSetData'
      produces:
      - application/json
      responses:
        "200":
          description: success
        "400":
          description: teamsize of elimination should not be 1
          schema:
            $ref: '#/definitions/response.ErrorReceiveDataFormatResponse'
        "500":
          description: internal db error / Get Player Set By Id / Update Player Set
            Name
          schema:
            $ref: '#/definitions/response.ErrorInternalErrorResponse'
      summary: Put player set name
      tags:
      - PlayerSet
  /playerset/preranking/{eliminationid}:
    patch:
      description: Put player set rank by elimination id
      parameters:
      - description: Elimination ID
        in: path
        name: eliminationid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
        "500":
          description: internal db error / Get Elimination Player Set Id Rank Order
            By Id / Update Player Set Rank
          schema:
            $ref: '#/definitions/response.ErrorInternalErrorResponse'
      summary: Put player set rank
      tags:
      - PlayerSet
...
swagger: "2.0"
...